requirements.txt
numpy
Pillow
cairosvg
torch
